refactor: improve stylesheets per component
