+++ src/components/AccountLogout.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import classNames from 'classnames'
 import { translate } from 'cozy-ui/react/I18n'
-import styles from '../styles/accountLoginForm'
+import styles from '../styles/accountLogout'
 
 export const AccountLogout = ({ t, deleting, onDelete }) => {
   return (
