+++ src/components/KonnectorInstall.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { translate } from 'cozy-ui/react/I18n'
-import styles from '../styles/accountConnection'
+import styles from '../styles/konnectorInstall'
 
 import AccountConnectionData from './AccountConnectionData'
 import AccountLoginForm from './AccountLoginForm'
