+++ src/styles/accountConnectionData.styl
@@ -1,0 +1,11 @@
+@import 'cozy-ui'
+
+:local // @stylint ignore
+    .col-account-connection-data
+        margin 0 (16/basefont)em
+
+        @media (max-width: (930/basefont)rem)
+            display none
+
+    .col-account-connection-data-access
+        padding-left 0
