+++ src/styles/accountLoginForm.styl
@@ -6,8 +6,7 @@
         display flex
         flex-wrap wrap
 
-    .coz-form-controls,
-    .col-account-form-delete
+    .coz-form-controls
         .coz-btn
             // Used to position progress bar
             flex 1 0 auto
@@ -23,9 +22,6 @@
         font-size (16/basefont)em
         height    (48/basefont)em
         width     100%
-
-    .col-account-form-delete .coz-btn
-        width 100%
 
     .col-account-form-success-buttons
         width 100%
