+++ src/styles/accountLogout.styl
@@ -1,0 +1,14 @@
+@import 'cozy-ui'
+
+:local // @stylint ignore
+    .col-account-form-delete
+        .coz-btn
+            width 100%
+            // Used to position progress bar
+            flex 1 0 auto
+
+            &:first-of-type
+                margin-left 0
+
+            &:last-of-type
+                margin-right 0
