+++ src/styles/konnectorEdit.styl
@@ -1,0 +1,12 @@
+@import 'cozy-ui'
+
+:local // @stylint ignore
+    .col-account-edit-content
+        margin-top 2em
+
+        .col-account-edit-tabs
+            .col-account-edit-tabpanel
+                padding-top 0
+
+            h4
+                margin-top 1.25rem
