+++ src/styles/konnectorInstall.styl
@@ -1,0 +1,26 @@
+@import 'cozy-ui'
+
+:local // @stylint ignore
+    .col-account-connection-content
+        display   flex
+        flex-wrap nowrap
+        width     (800/basefont)em
+
+    .col-account-connection-security svg
+        max-height   (48/basefont)em
+        max-width    (48/basefont)em
+        float        left
+        margin-right (16/basefont)em
+
+    .col-account-connection-form
+        margin 0 (16/basefont)em
+        flex 1 0 (460/basefont)em
+        max-width (460/basefont)em
+
+
+    @media (max-width: (930/basefont)rem)
+        .col-account-connection-content
+            width   100%
+
+        .col-account-connection-form
+            flex    1 0
